Adding more tests on BaseExceptions; breaking out test problems to common file
